00-framework/requirements.txt
pysftp>=0.2.9
python-dotenv